Move seed backup page to seedbackup package
